Implement InputCountainer component
Implement the InputContainer component according to test.
And Refactor the Page component.
