+++ src/Page.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { render, fireEvent } from '@testing-library/react';
+import { render } from '@testing-library/react';
 
 import { useSelector, useDispatch } from 'react-redux';
 
@@ -9,9 +9,6 @@ import Page from './Page';
 jest.mock('react-redux');
 
 test('Page', () => {
-  const handleChangeTitle = jest.fn();
-  const handleClickAddTask = jest.fn();
-
   const dispatch = jest.fn();
   useDispatch.mockImplementation(() => dispatch);
   useSelector.mockImplementation((selector) => selector({
@@ -22,17 +19,9 @@ test('Page', () => {
   }));
 
   const { getByText } = render((
-    <Page
-      taskTitle=""
-      onChangeTitle={handleChangeTitle}
-      onClickAddTask={handleClickAddTask}
-    />
+    <Page />
   ));
 
   expect(getByText(/task-1/)).not.toBeNull();
   expect(getByText(/task-2/)).not.toBeNull();
-
-  fireEvent.click(getByText('추가'));
-
-  expect(handleClickAddTask).toBeCalled();
 });
